requirements.txt
requests==2.25.1
Django==3.1.5
djangorestframework==3.12.4
aniso8601==9.0.1
pysendpulse==0.1.2
python-dateutil==2.8.1
python3-memcached==1.51
mysqlclient==2.0.3
python-dotenv==0.18.0
